Split updater code from contract into own module
