+++ include/updater/updater_config.hpp
@@ -25,10 +25,8 @@ SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
 */
 
-#ifndef CONTRACTOR_OPTIONS_HPP
-#define CONTRACTOR_OPTIONS_HPP
-
-#include "updater/updater_config.hpp"
+#ifndef OSRM_UPDATER_UPDATER_CONFIG_HPP
+#define OSRM_UPDATER_UPDATER_CONFIG_HPP
 
 #include <boost/filesystem/path.hpp>
 
@@ -36,45 +34,47 @@ SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
 namespace osrm
 {
-namespace contractor
+namespace updater
 {
 
-struct ContractorConfig
+struct UpdaterConfig final
 {
-    ContractorConfig() : requested_num_threads(0) {}
-
     // Infer the output names from the path of the .osrm file
     void UseDefaultOutputNames()
     {
-        level_output_path = osrm_input_path.string() + ".level";
-        core_output_path = osrm_input_path.string() + ".core";
-        graph_output_path = osrm_input_path.string() + ".hsgr";
-        node_file_path = osrm_input_path.string() + ".enw";
-        updater_config.osrm_input_path = osrm_input_path;
-        updater_config.UseDefaultOutputNames();
+        edge_based_graph_path = osrm_input_path.string() + ".ebg";
+        edge_segment_lookup_path = osrm_input_path.string() + ".edge_segment_lookup";
+        turn_weight_penalties_path = osrm_input_path.string() + ".turn_weight_penalties";
+        turn_duration_penalties_path = osrm_input_path.string() + ".turn_duration_penalties";
+        turn_penalties_index_path = osrm_input_path.string() + ".turn_penalties_index";
+        node_based_graph_path = osrm_input_path.string() + ".nodes";
+        geometry_path = osrm_input_path.string() + ".geometry";
+        rtree_leaf_path = osrm_input_path.string() + ".fileIndex";
+        datasource_names_path = osrm_input_path.string() + ".datasource_names";
+        datasource_indexes_path = osrm_input_path.string() + ".datasource_indexes";
+        profile_properties_path = osrm_input_path.string() + ".properties";
     }
-
-    updater::UpdaterConfig updater_config;
 
     boost::filesystem::path osrm_input_path;
 
-    std::string level_output_path;
-    std::string core_output_path;
-    std::string graph_output_path;
+    std::string edge_based_graph_path;
 
-    std::string node_file_path;
+    std::string edge_segment_lookup_path;
+    std::string turn_weight_penalties_path;
+    std::string turn_duration_penalties_path;
+    std::string turn_penalties_index_path;
+    std::string node_based_graph_path;
+    std::string geometry_path;
+    std::string rtree_leaf_path;
 
-    bool use_cached_priority;
+    double log_edge_updates_factor;
 
-    unsigned requested_num_threads;
-
-    // A percentage of vertices that will be contracted for the hierarchy.
-    // Offers a trade-off between preprocessing and query time.
-    // The remaining vertices form the core of the hierarchy
-    //(e.g. 0.8 contracts 80 percent of the hierarchy, leaving a core of 20%)
-    double core_factor;
+    std::vector<std::string> segment_speed_lookup_paths;
+    std::vector<std::string> turn_penalty_lookup_paths;
+    std::string datasource_indexes_path;
+    std::string datasource_names_path;
+    std::string profile_properties_path;
 };
-
 }
 }
